chore(site): generate pages json data at build
